Use the IPC functionality of mpv to monitor video status
Instead of loading mpv via `GLib.spawn_async` and leaving it to its own
devices create and connect to the mpv IPC socket. A background thread
does blocking reads on the mpv socket notifies the UI for changes. This
also allows to provide better feedback to the user about the status of
the current video being loaded. In addition, if a video is already
playing, loading of a new video will be done through the existing mpv
instance instead of spawning a brand new process.